requirements.txt
gradio
torch
torchvision
transformers
opencv-python
numpy
pandas
requests
python-dotenv
#facenet-pytorch 
#ffmpeg-python